pyproject.toml
[project]
name = "ops-library-testing"
version = "0.1.0"
description = "Testing environment for ops-library Ansible roles and services"
requires-python = ">=3.8"
dependencies = [
    "ansible-core>=2.15",
    "ansible-lint>=6.0",
    "yamllint>=1.32",
    "pytest>=7.0",
    "pyyaml>=6.0",
    "jinja2>=3.0",
    "markupsafe>=2.0",
    "pathspec>=0.11",
    "molecule>=5.0",
    "pre-commit>=3.5.0",
]

[project.optional-dependencies]
docker = [
    "molecule-docker>=2.0",
]
vagrant = [
    "molecule-vagrant>=2.0",
    "python-vagrant>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"